# service/aiServeWeaveGateway/openapi.yaml
openapi: 3.0.3
info:
  title: AIServeWeave Gateway API
  description: |
    aiServeWeaveGateway 对外暴露的两组 HTTP 接口：

    - **公开推理前门**（默认监听地址，需要租户 API Key）：OpenAI 兼容的
      chat/embeddings/responses/models，以及工作流 Job 的提交、状态、SSE 事件流、
      取消与产物下载。
    - **只读运维接口**（`-admin-addr` 单独配置时才挂载，使用部署密钥而非租户 Key）：
      机群清单、工作流目录、按租户的 job 列表与路由生效状态，供控制面聚合展示，不做任何写操作。

    源码位置：`service/aiServeWeaveGateway/httpapi/`（公开前门）与
    `service/aiServeWeaveGateway/adminapi/`（运维接口）。
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: 公开推理前门（默认端口，可通过 -addr 配置）
  - url: http://localhost:8081
    description: 只读运维接口（通过 -admin-addr 单独配置，未配置则不启动）
...
security:
  - bearerAuth: []
paths:
...
  /internal/v1/routes:
    get:
      tags: [admin]
      summary: 本副本实际应用的路由版本与最近同步状态
      description: 仅在运维监听器提供；文件模式不会被视为控制面版本确认。错误为固定代号，不包含配置或凭据。
      operationId: getAppliedRoutes
      security:
        - adminToken: []
      responses:
        "200":
          description: 路由生效状态
          content:
            application/json:
              schema:
                type: object
                required: [replica_id, generated_at, mode, revision, digest, applied_at, checked_at]
                properties:
                  replica_id: {type: string}
                  generated_at: {type: string, format: date-time}
                  mode: {type: string, enum: [file, controlplane]}
                  revision: {type: integer, format: int64, minimum: 0}
                  digest: {type: string}
                  applied_at: {type: string, format: date-time}
                  checked_at: {type: string, format: date-time}
                  error: {type: string}
        "401":
          $ref: "#/components/responses/AdminError"
  /internal/v1/nodes:
    get:
      tags: [admin]
      summary: 本副本当前看到的机群清单
      operationId: getFleetSnapshot
      security:
        - adminToken: []
      responses:
        "200":
          description: 机群快照
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Replica"
        "401":
          $ref: "#/components/responses/AdminError"
  /internal/v1/workflows:
    get:
      tags: [admin]
      summary: 本副本注册的工作流模板目录（未配置模板目录时不挂载此路由）
      operationId: getWorkflowCatalog
      security:
        - adminToken: []
      responses:
        "200":
          description: 模板目录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateCatalog"
        "401":
          $ref: "#/components/responses/AdminError"
  /internal/v1/jobs:
    get:
      tags: [admin]
      summary: 本副本上某一个租户的运行列表（未启用前门时不挂载此路由）
      description: tenant_id 为必填查询参数；缺省时返回 400，而不是返回全部租户的运行。
      operationId: getTenantJobs
      security:
        - adminToken: []
      parameters:
        - name: tenant_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 该租户在本副本上的运行
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPage"
        "400":
          $ref: "#/components/responses/AdminError"
        "401":
          $ref: "#/components/responses/AdminError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "租户 API Key（Authorization: Bearer <key>）。未配置任何鉴权方式的部署会放行一切请求。"
    adminToken:
      type: http
      scheme: bearer
      description: 部署级运维密钥（AISW_GATEWAY_ADMIN_TOKEN），与租户 API Key 无关。
...
  responses:
...
    AdminError:
      description: 运维接口的错误体
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
  schemas:
...
    Replica:
      type: object
      properties:
        replica_id: { type: string }
        generated_at: { type: string, format: date-time }
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/Node"
    Node:
      type: object
      properties:
        node_id: { type: string }
        agent_version: { type: string }
        live: { type: boolean }
        draining: { type: boolean }
        last_heartbeat: { type: string, format: date-time }
        labels:
          type: object
          additionalProperties: { type: string }
        resources:
          type: object
          properties:
            cpu_cores: { type: integer }
            memory_bytes: { type: integer, format: int64 }
            gpu_count: { type: integer }
            gpu_memory_bytes: { type: integer, format: int64 }
            os: { type: string }
            arch: { type: string }
        inflight_requests: { type: integer }
        idle_slots:
          type: object
          additionalProperties: { type: integer }
          description: "槽位 class（inference/bulk）到空闲数量"
        declared_runtime_ids:
          type: array
          items: { type: string }
        runtimes:
          type: array
          items:
            $ref: "#/components/schemas/Runtime"
    Runtime:
      type: object
      properties:
        id: { type: string }
        kind: { type: string }
        base_url: { type: string }
        state: { type: string }
        version: { type: string }
        identity_verified: { type: boolean }
        latency_millis: { type: integer, format: int64 }
        checked_at: { type: string, format: date-time }
        error_summary: { type: string }
        capabilities:
          type: object
          additionalProperties: { type: string }
        models:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              capabilities:
                type: object
                additionalProperties: { type: string }
        warnings:
          type: array
          items: { type: string }
        degraded:
          type: array
          items: { type: string }
        discovered_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    TemplateCatalog:
      type: object
      properties:
        replica_id: { type: string }
        generated_at: { type: string, format: date-time }
        templates:
          type: array
          items:
            $ref: "#/components/schemas/Template"
    Template:
      type: object
      properties:
        id: { type: string }
        description: { type: string }
        valid: { type: boolean }
        validation_error: { type: string }
        inputs:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              type: { type: string }
              required: { type: boolean }
              default: {}
              max_length: { type: integer }
              min: { type: number }
              max: { type: number }
    JobPage:
      type: object
      properties:
        replica_id: { type: string }
        generated_at: { type: string, format: date-time }
        tenant_id: { type: string }
        truncated: { type: boolean }
        jobs:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              workflow_id: { type: string }
              state: { type: string }
              queue_position: { type: integer }
              error_summary: { type: string }
              created_at: { type: string, format: date-time }
              updated_at: { type: string, format: date-time }
              artifact_ids:
                type: array
                items: { type: string }